backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
celery==5.3.4
redis==5.0.1
openai==1.3.5
replicate==0.15.4
elevenlabs==0.2.26
boto3==1.29.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
moviepy==1.0.3
python-multipart==0.0.6
gunicorn==21.2.0